requirements.txt
fastapi==0.115.0
starlette==0.38.5
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-core==2.23.4
pymongo==4.8.0
requests==2.32.3
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
